requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-storage-blob
azure-identity
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
dnstwist==20230413
frozenlist==1.3.3
future==0.18.3
geoip2==4.6.0
idna==3.4
logger==1.4
maxminddb==2.2.0
multidict==6.0.4
ppdeep==20200505
#py-tlsh==4.7.2
python-whois==0.8.0
requests==2.29.0
termcolor==2.3.0
tld==0.13
urllib3==1.26.15
websocket-client==1.5.1
yarl==1.9.2
